Refactor attributes model package structure
