+++ example/src/test/java/org/wordpress/android/fluxc/wc/attributes/WCGlobalAttributeMapperTest.kt
@@ -4,7 +4,6 @@ import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
 import com.yarolegovich.wellsql.WellSql
 import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.fail
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -13,16 +12,17 @@ import org.robolectric.RuntimeEnvironment
 import org.robolectric.annotation.Config
 import org.wordpress.android.fluxc.SingleStoreWellSqlConfigForTests
 import org.wordpress.android.fluxc.model.SiteModel
-import org.wordpress.android.fluxc.model.product.attributes.WCGlobalAttributeMapper
-import org.wordpress.android.fluxc.model.product.attributes.terms.WCAttributeTermModel
+import org.wordpress.android.fluxc.model.attribute.WCGlobalAttributeMapper
+import org.wordpress.android.fluxc.model.attribute.terms.WCAttributeTermModel
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooPayload
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.product.attributes.ProductAttributeRestClient
 import org.wordpress.android.fluxc.persistence.WellSqlConfig
 import org.wordpress.android.fluxc.test
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.attributeCreateResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.attributeTermsFullListResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.attributesFullListResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.stubSite
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.attributeCreateResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.attributeTermsFullListResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.attributesFullListResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.stubSite
+import kotlin.test.fail
 
 @Config(manifest = Config.NONE)
 @RunWith(RobolectricTestRunner::class)
