+++ example/src/test/java/org/wordpress/android/fluxc/wc/attributes/WCGlobalAttributeStoreTest.kt
@@ -14,25 +14,25 @@ import org.robolectric.RuntimeEnvironment
 import org.robolectric.annotation.Config
 import org.wordpress.android.fluxc.SingleStoreWellSqlConfigForTests
 import org.wordpress.android.fluxc.model.SiteModel
-import org.wordpress.android.fluxc.model.product.attributes.WCGlobalAttributeMapper
-import org.wordpress.android.fluxc.model.product.attributes.WCGlobalAttributeModel
-import org.wordpress.android.fluxc.model.product.attributes.terms.WCAttributeTermModel
+import org.wordpress.android.fluxc.model.attribute.WCGlobalAttributeMapper
+import org.wordpress.android.fluxc.model.attribute.WCGlobalAttributeModel
+import org.wordpress.android.fluxc.model.attribute.terms.WCAttributeTermModel
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooPayload
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.product.attributes.ProductAttributeRestClient
 import org.wordpress.android.fluxc.persistence.WellSqlConfig
 import org.wordpress.android.fluxc.store.WCGlobalAttributeStore
 import org.wordpress.android.fluxc.test
 import org.wordpress.android.fluxc.tools.initCoroutineEngine
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.attributeCreateResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.attributeDeleteResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.attributeTermsFullListResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.attributeUpdateResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.attributesFullListResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.parsedAttributesList
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.parsedCreateAttributeResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.parsedDeleteAttributeResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.parsedUpdateAttributeResponse
-import org.wordpress.android.fluxc.wc.attributes.WCGlobalAttributesTestFixtures.stubSite
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.attributeCreateResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.attributeDeleteResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.attributeTermsFullListResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.attributeUpdateResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.attributesFullListResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.parsedAttributesList
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.parsedCreateAttributeResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.parsedDeleteAttributeResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.parsedUpdateAttributeResponse
+import org.wordpress.android.fluxc.wc.attributes.WCProductAttributesTestFixtures.stubSite
 
 @Config(manifest = Config.NONE)
 @RunWith(RobolectricTestRunner::class)
