+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/model/attribute/WCGlobalAttributeMapper.kt
@@ -1,7 +1,7 @@
-package org.wordpress.android.fluxc.model.product.attributes
+package org.wordpress.android.fluxc.model.attribute
 
 import org.wordpress.android.fluxc.model.SiteModel
-import org.wordpress.android.fluxc.model.product.attributes.terms.WCAttributeTermModel
+import org.wordpress.android.fluxc.model.attribute.terms.WCAttributeTermModel
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.product.attributes.AttributeApiResponse
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.product.attributes.ProductAttributeRestClient
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.product.attributes.terms.AttributeTermApiResponse
