+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/model/attribute/terms/WCAttributeTermModel.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.fluxc.model.product.attributes.terms
+package org.wordpress.android.fluxc.model.attribute.terms
 
 import com.yarolegovich.wellsql.core.Identifiable
 import com.yarolegovich.wellsql.core.annotation.Column
